Support ingest of data not aligned to tile grid.
In order to support a variety of configurations, ingest needs to handle data that isn't pre-aligned to the tiling grid. This update gives data curators a way to specify the tile size and it handles framing data (with fill values) in order to get input data precisely aligned with the tile grid. However, source data must still have a matching pixel size, and have extents that are "on" the pixel grid. In other words, the upper left extents of a raster must be a multiple of the pixel size. An exception is generated during tiling when this expectaton is not met.

This change refactors the scene and band classes to facilitate better automated testing. Test data includes ESPA metadata (but not ESPA rasters) and sample raster data generated using GDAL. The decision to use sample raster data was made in order to keep the repository size small. Sample data from ESPA could be several megabytes; adding those to the repository lacked appeal.

The ingest script itself was changed slightly. Unused parameters (for the keyspace and table) were removed, and additional parameters for the tile size and expected pixel size were added.