Handle offset in snapping function, simplify offset during tiling.
Instead of using command line parameters, we can inspect the raster data to determine the amount of offset between an image and the projection's implicit raster grid. For example, an EPSG:5070 image projected at 30 meters, has an implied grid unit every 30 meters. However, if the upper left pixel coordinate is *not* a factor of 30 then the grid is "offset" in a direction. This is determined by taking the upper left coordinate modulo the pixel size.

Taking this approach does mean that a curator needs to know exactly what they are doing, otherwise they could ingest data that is not in harmony into the same table. This potential problem will be documented and handled later when more metadata about each ARD dataset is added to the DB.

The snap function has been updated to handle an offset too so that the tile containing data for an arbitrary pixel can be found.

commit a5c9b173e34a053ba7ae6f24b1c2ade3ef959a58
Author: Jonathan Morton <jmorton@usgs.gov>
Date:   Mon Sep 21 11:42:51 2015 -0500

    Add additional comments and remove unused script parameters.

commit 6af84eb95ce69bb0b034afef8f3244135fdf78b3
Author: Jonathan Morton <jmorton@usgs.gov>
Date:   Mon Sep 21 10:43:26 2015 -0500

    Add more comprehensive tests, fix spap to upper left feature

commit c3eb623ab6d1e93412f877608cbbe949f03b8218
Author: Jonathan Morton <jmorton@usgs.gov>
Date:   Mon Sep 21 09:02:52 2015 -0500

    In progress commit